move Pledge parsing/handling logic to service layer
