+++ test/lib/parse/url.js
@@ -3,75 +3,75 @@ import { expect } from 'chai'
 import isValidUrl, { isBotFromUserAgent } from 'lib/parse/url'
 
 describe('isValidUrl', () => {
-  def('url', () => '/go_for_a_run/by/2:30pm')
-  subject(() => isValidUrl({ url: $url }))
+  def('url', () => '/go_for_a_run/by/2:30pm');
+  subject(() => isValidUrl({ url: $url }));
 
   it('allows any url that has no invalid characters', () => {
-    expect($subject).to.have.length(0)
+    expect($subject.valid).to.be.true;
   })
 
   context('when the url contains a % sign', () => {
-    def('url', () => '/space%20here')
+    def('url', () => '/space%20here');
 
     it('rejects the % sign', () => {
-      expect($subject).to.have.length(1)
+      expect($subject.valid).to.be.false;
     })
   })
 
   context('when the url contains an apostrophe', () => {
-    def('url', () => '/for-it\'s-test')
+    def('url', () => '/for-it\'s-test');
 
     it('rejects the apostrophe', () => {
-      expect($subject).to.have.length(1)
+      expect($subject.valid).to.be.false;
     })
   })
 
   context('when the url has a file extension', () => {
-    def('url', () => '/awekjad.txt')
+    def('url', () => '/awekjad.txt');
 
     it('rejects the file extension', () => {
-      expect($subject).to.have.length(1)
+      expect($subject.valid).to.be.false;
     })
   })
 
   context('when the url ends in abby', () => {
-    def('url', () => '/test/this/thing/for/abby')
+    def('url', () => '/test/this/thing/for/abby');
 
     it('does not reject the url', () => {
-      expect($subject).to.have.length(0)
+      expect($subject.valid).to.be.true;
     })
   })
 
   context('when the url ends in /by', () => {
-    def('url', () => '/test/this/thing/by')
+    def('url', () => '/test/this/thing/by');
 
     it('rejects the /by', () => {
-      expect($subject).to.have.length(1)
+      expect($subject.valid).to.be.false;
     })
   })
 
   context('when the url repeats the username and domain', () => {
-    def('url', () => '/tester.commits.to/test/the/things')
+    def('url', () => '/tester.commits.to/test/the/things');
 
     it('rejects the url', () => {
-      expect($subject).to.have.length(1)
+      expect($subject.valid).to.be.false;
     })
   })
 })
 
 describe('isQueryString', () => {
   def('url', () => '/go_running?by=2:30pm')
-  subject(() => isValidUrl({ url: $url }))
+  subject(() => isValidUrl({ url: $url }));
 
   it('rejects any url that has a "valid" querystring', () => {
-    expect($subject).to.have.length(1)
+    expect($subject.valid).to.be.false;
   })
 
   context('when the "?" is followed by a "/"', () => {
-    def('url', () => 'go_running?/by/2:30pm')
+    def('url', () => 'go_running?/by/2:30pm');
 
     it('does not consider the url to have a querystring', () => {
-      expect($subject).to.have.length(0)
+      expect($subject.valid).to.be.true;
     })
   })
 })
